Add "CartDatabase" class for save products in database NOTE: 1. Add (syncCartAddItem), (syncCartRemoveItem), (syncCartGetItems) methods into CartDatabase class to avoid repeatation
