+++ frontend/JS/cart.js
@@ -2,6 +2,7 @@ import accountHandler from "./utils/account.handler.js";
 import isTokenExist from "./utils/check.token.js";
 import logoutHandler from "./utils/logout.handler.js";
 import applyTheme from "./utils/theme.js";
+import cartDatabase from "./utils/cart.database.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   isTokenExist();
@@ -131,7 +132,7 @@ document.addEventListener("DOMContentLoaded", () => {
         item.qty -= 1;
       }
 
-      syncCartAddItem(id, item.qty);
+      cartDatabase.syncCartAddItem(id, item.qty);
 
       // Remove if qty <= 0
       const newCart = cart.filter((p) => p.qty > 0);
@@ -147,7 +148,7 @@ document.addEventListener("DOMContentLoaded", () => {
       const newCart = cart.filter((p) => p.id != id);
       saveCart(newCart);
 
-      syncCartRemoveItem(id);
+      cartDatabase.syncCartRemoveItem(id);
 
       renderCart();
       updateCartCount();
@@ -172,31 +173,3 @@ document.addEventListener("DOMContentLoaded", () => {
   renderCart();
   updateCartCount();
 });
-
-// Cart Req To Backend
-async function syncCartAddItem(productId, quantity) {
-  const token = localStorage.getItem("token");
-  const userId = JSON.parse(atob(token.split(".")[1])).id;
-
-  await fetch("http://localhost:3000/cart/add", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ userId, productId, quantity }),
-  });
-}
-async function syncCartRemoveItem(productId) {
-  const token = localStorage.getItem("token");
-  const userId = JSON.parse(atob(token.split(".")[1])).id;
-
-  await fetch("http://localhost:3000/cart/remove", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ userId, productId }),
-  });
-}
