+++ frontend/JS/main.js
@@ -3,6 +3,7 @@ import applyTheme from "./utils/theme.js";
 import logoutHandler from "./utils/logout.handler.js";
 import sampleProducts from "./utils/sample.products.js";
 import accountHandler from "./utils/account.handler.js";
+import cartDatabase from "./utils/cart.database.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
   const productGrid = document.getElementById("product-grid");
@@ -95,7 +96,7 @@ function addToCart(product) {
   const existing = cart.find((item) => item.id == product.id);
 
   const newQty = existing ? existing.qty : 1;
-  syncCartAddItem(product.id, newQty);
+  cartDatabase.syncCartAddItem(product.id, newQty);
 
   if (existing) {
     existing.qty += 1;
@@ -114,7 +115,7 @@ function increaseQty(id) {
   let cart = getCart();
   let item = cart.find((p) => p.id == id);
   item.qty += 1;
-  syncCartAddItem(id, item.qty);
+  cartDatabase.syncCartAddItem(id, item.qty);
   saveCart(cart);
   updateCartCount();
   return item.qty;
@@ -126,13 +127,13 @@ function decreaseQty(id) {
 
   if (item.qty > 1) {
     item.qty -= 1;
-    syncCartAddItem(id, item.qty);
+    cartDatabase.syncCartAddItem(id, item.qty);
     saveCart(cart);
     updateCartCount();
     return item.qty;
   } else {
     cart = cart.filter((p) => p.id != id);
-    syncCartRemoveItem(id);
+    cartDatabase.syncCartRemoveItem(id);
     saveCart(cart);
     updateCartCount();
     return 0;
@@ -200,32 +201,4 @@ function updateCartCount() {
   if (countElement) {
     countElement.textContent = count;
   }
-}
-
-// Cart Req To Backend
-async function syncCartAddItem(productId, quantity) {
-  const token = localStorage.getItem("token");
-  const userId = JSON.parse(atob(token.split(".")[1])).id;
-
-  await fetch("http://localhost:3000/cart/add", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ userId, productId, quantity }),
-  });
-}
-async function syncCartRemoveItem(productId) {
-  const token = localStorage.getItem("token");
-  const userId = JSON.parse(atob(token.split(".")[1])).id;
-
-  await fetch("http://localhost:3000/cart/remove", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ userId, productId }),
-  });
 }
